incrementalmerkletree
Module bridgetree

A space-efficient implementation of the Tree interface.

In this module, the term “ommer” is used as a gender-neutral term for the sibling of a parent node in a binary tree.

Structs

AuthFragment

Each AuthFragment stores part of the authentication path for the leaf at a particular position. Successive fragments may be concatenated to produce the authentication path up to one less than the maximum altitude of the Merkle frontier corresponding to the leaf at the specified position. Then, the authentication path may be completed by hashing with empty roots.

BridgeTree

Checkpoint

Frontier

A possibly-empty Merkle frontier. Used when the full functionality of a Merkle bridge is not necessary.

MerkleBridge

NonEmptyFrontier

A [NonEmptyFrontier] is a reduced representation of a Merkle tree, having either one or two leaf values, and then a set of hashes produced by the reduction of previously appended leaf values.

Enums

BridgeTreeError

Errors that can appear when validating the internal consistency of a [MerkleBridge] value when constructing a bridge from its constituent parts.

FrontierError

Leaf

A set of leaves of a Merkle tree.