Struct incrementalmerkletree::bridgetree::NonEmptyFrontier

pub struct NonEmptyFrontier<H> { /* private fields */ }

A [NonEmptyFrontier] is a reduced representation of a Merkle tree, having either one or two leaf values, and then a set of hashes produced by the reduction of previously appended leaf values.

Implementations

impl<H> NonEmptyFrontier<H>

pub fn new(value: H) -> Self

Constructs a new frontier with the specified value at position 0.

pub fn from_parts(
position: Position,
leaf: Leaf<H>,
ommers: Vec<H>
) -> Result<Self, FrontierError>

pub fn max_altitude(&self) -> Altitude

Returns the altitude of the highest ommer in the frontier.

pub fn position(&self) -> Position

Returns the position of the most recently appended leaf.

pub fn size(&self) -> usize

Returns the number of leaves that have been appended to this frontier.

pub fn leaf(&self) -> &Leaf<H>

pub fn ommers(&self) -> &[H]

pub fn leaf_value(&self) -> &H

Returns the value of the most recently appended leaf.

impl<H: Hashable + Clone> NonEmptyFrontier<H>

pub fn current_leaf(&self) -> &H

pub fn append(&mut self, value: H)

of two nodes is full (if the current leaf before the append is a Leaf::Right) then recompute the ommers by hashing together full subtrees until an empty ommer slot is found.

pub fn root(&self) -> H

Generate the root of the Merkle tree by hashing against empty branches.

pub fn witness(&self, sibling_altitude: Altitude) -> Option<H>

If the tree is full to the specified altitude, return the data required to witness a sibling at that altitude.

pub fn witness_incomplete(&self, altitude: Altitude) -> Option<H>

If the tree is not full, generate the root of the incomplete subtree by hashing with empty branches