Trait incrementalmerkletree::Frontier

pub trait Frontier<H> {
    // Required methods
    fn append(&mut self, value: &H) -> bool;
    fn root(&self) -> H;
}

A possibly-empty incremental Merkle frontier.

Required Methods

fn append(&mut self, value: &H) -> bool

Appends a new value to the frontier at the next available slot. Returns true if successful and false if the frontier would exceed the maximum allowed depth.

fn root(&self) -> H

Obtains the current root of this Merkle frontier by hashing against empty nodes up to the maximum height of the pruned tree that the frontier represents.