incrementalmerkletree
Trait Hashable

pub trait Hashable: Sized {
    // Required methods
    fn empty_leaf() -> Self;
    fn combine(level: Altitude, a: &Self, b: &Self) -> Self;

    // Provided method
    fn empty_root(level: Altitude) -> Self { ... }
}

A trait describing the operations that make a value suitable for inclusion in an incremental merkle tree.

Required Methods

fn empty_leaf() -> Self

fn combine(level: Altitude, a: &Self, b: &Self) -> Self

Provided Methods

fn empty_root(level: Altitude) -> Self

Object Safety

This trait is not object safe.