incrementalmerkletree
Trait Tree

pub trait Tree<H> {
    // Required methods
    fn append(&mut self, value: &H) -> bool;
    fn current_position(&self) -> Option<Position>;
    fn current_leaf(&self) -> Option<&H>;
    fn get_witnessed_leaf(&self, position: Position) -> Option<&H>;
    fn witness(&mut self) -> Option<Position>;
    fn witnessed_positions(&self) -> BTreeSet<Position>;
    fn root(&self, checkpoint_depth: usize) -> Option<H>;
    fn authentication_path(
        &self,
        position: Position,
        as_of_root: &H,
    ) -> Option<Vec<H>>;
    fn remove_witness(&mut self, position: Position) -> bool;
    fn checkpoint(&mut self);
    fn rewind(&mut self) -> bool;
    fn garbage_collect(&mut self);
}

A Merkle tree that supports incremental appends, witnessing of leaf nodes, checkpoints and rollbacks.

Required Methods

fn append(&mut self, value: &H) -> bool

Appends a new value to the tree at the next available slot. Returns true if successful and false if the tree would exceed the maximum allowed depth.

fn current_position(&self) -> Option<Position>

Returns the most recently appended leaf value.

fn current_leaf(&self) -> Option<&H>

Returns the most recently appended leaf value.

fn get_witnessed_leaf(&self, position: Position) -> Option<&H>

Returns the leaf at the specified position if the tree can produce an authentication path for it.

fn witness(&mut self) -> Option<Position>

Marks the current leaf as one for which we’re interested in producing an authentication path. Returns an optional value containing the current position if successful or if the current value was already marked, or None if the tree is empty.

fn witnessed_positions(&self) -> BTreeSet<Position>

Return a set of all the positions for which we have witnessed.

fn root(&self, checkpoint_depth: usize) -> Option<H>

Obtains the root of the Merkle tree at the specified checkpoint depth by hashing against empty nodes up to the maximum height of the tree. Returns None if there are not enough checkpoints available to reach the requested checkpoint depth.

fn authentication_path( &self, position: Position, as_of_root: &H, ) -> Option<Vec<H>>

Obtains an authentication path to the value at the specified position, as of the tree state corresponding to the given root. Returns None if there is no available authentication path to that position or if the root does not correspond to a checkpointed root of the tree.

fn remove_witness(&mut self, position: Position) -> bool

Marks the value at the specified position as a value we’re no longer interested in maintaining a witness for. Returns true if successful and false if we were already not maintaining a witness at this position.

fn checkpoint(&mut self)

Creates a new checkpoint for the current tree state. It is valid to have multiple checkpoints for the same tree state, and each rewind call will remove a single checkpoint.

fn rewind(&mut self) -> bool

Rewinds the tree state to the previous checkpoint, and then removes that checkpoint record. If there are multiple checkpoints at a given tree state, the tree state will not be altered until all checkpoints at that tree state have been removed using rewind. This function return false and leave the tree unmodified if no checkpoints exist.

fn garbage_collect(&mut self)

Remove state from the tree that no longer needs to be maintained because it is associated with checkpoints or witnesses that have been removed from the tree at positions deeper than those reachable by calls to rewind. It is always safe to implement this as a no-op operation