Struct incrementalmerkletree::bridgetree::BridgeTree

pub struct BridgeTree<H: Ord, const DEPTH: u8> { /* private fields */ }

Implementations

impl<H: Ord, const DEPTH: u8> BridgeTree<H, DEPTH>

pub fn prior_bridges(&self) -> &[MerkleBridge<H>]

Returns the prior bridges that make up this tree

pub fn current_bridge(&self) -> &Option<MerkleBridge<H>>

Returns the current bridge at the tip of this tree

pub fn witnessed_indices(&self) -> &BTreeMap<Position, usize>

pub fn checkpoints(&self) -> &[Checkpoint]

Returns the checkpoints to which this tree may be rewound.

pub fn max_checkpoints(&self) -> usize

Returns the maximum number of checkpoints that will be maintained by the data structure. When this number of checkpoints is exceeded, the oldest checkpoints are discarded when creating new checkpoints.

pub fn frontier(&self) -> Option<&NonEmptyFrontier<H>>

impl<H: Hashable + Ord + Clone, const DEPTH: u8> BridgeTree<H, DEPTH>

pub fn new(max_checkpoints: usize) -> Self

pub fn from_frontier(
max_checkpoints: usize,
frontier: NonEmptyFrontier<H>
) -> Self

Construct a new BridgeTree that will start recording changes from the state of the specified frontier.

pub fn from_parts(
prior_bridges: Vec<MerkleBridge<H>>,
current_bridge: Option<MerkleBridge<H>>,
saved: BTreeMap<Position, usize>,
checkpoints: Vec<Checkpoint>,
max_checkpoints: usize
) -> Result<Self, BridgeTreeError>

pub fn check_consistency(&self) -> Result<(), BridgeTreeError>